Reorganize AI test utils + docs
