+++ src/@types/ProtectConditionFunc.ts
@@ -1,0 +1,3 @@
+import type { MoveId } from "#enums/move-id";
+
+export type ProtectConditionFunc = (moveId: MoveId) => boolean;
