+++ src/constants/ability-constants.ts
@@ -20,6 +20,7 @@ export const DETRIMENTAL_ABILITIES: Readonly<AbilityId[]> = Object.freeze([
   AbilityId.SLOW_START,
 ]);
 
+/** Abilities that grant an immunity to Electric-type moves */
 export const ELECTRIC_IMMUNE_ABILITIES: Readonly<AbilityId[]> = Object.freeze([
   AbilityId.VOLT_ABSORB,
   AbilityId.LIGHTNING_ROD,
