+++ src/data/moves/move-attrs/happy-hour-attr.ts
@@ -6,6 +6,11 @@ import type { Pokemon } from "#app/field/pokemon";
 import { ArenaTagRelativeSide } from "#enums/arena-tag-relative-side";
 import { ArenaTagType } from "#enums/arena-tag-type";
 
+/**
+ * Attribute to apply the effect of {@link https://bulbapedia.bulbagarden.net/wiki/Happy_Hour_(move) | Happy Hour}.
+ * Doubles the money given to the Player at the end of the current battle.
+ * @extends AddArenaTagAttr
+ */
 export class HappyHourAttr extends AddArenaTagAttr {
   constructor() {
     super(ArenaTagType.HAPPY_HOUR, ArenaTagRelativeSide.USER, { failOnOverlap: true });
@@ -13,7 +18,7 @@ export class HappyHourAttr extends AddArenaTagAttr {
 
   /**
    * Grants a {@link BAD_MOVE_PENALTY | major penalty} in all game states.
-   * As it stands, this attribute just gives money to the Player
+   * Currently, this attribute just gives money to the Player
    * no matter who uses it.
    */
   public override getEffectScore(_user: EnemyPokemon, _target: Pokemon, _move: Move): number {
