+++ test/ai/move-effect-scores/ability-copy.test.ts
@@ -2,7 +2,7 @@ import { DETRIMENTAL_ABILITIES } from "#app/constants/ability-constants";
 import { AbilityId } from "#enums/ability-id";
 import { MoveId } from "#enums/move-id";
 import { SpeciesId } from "#enums/species-id";
-import { revealAllAbilities } from "#test/ai/utils/enemy-command-utils";
+import { revealAllAbilities } from "#test/test-utils/enemy-command-utils";
 import { GameManager } from "#test/test-utils/gameManager";
 import { afterEach, beforeAll, beforeEach, describe, expect, it } from "vitest";
