+++ test/test-utils/enemy-command-utils.ts
